requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
altair==5.5.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2024.6.2
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
distlib==0.3.8
Django==4.2.11
django-auditlog==3.0.0
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.14.0
djangorestframework==3.15.1
drf-excel==2.4.0
drf-flex-fields==1.0.2
et-xmlfile==1.1.0
filelock==3.15.1
Flask==2.3.2
Flask-Cors==4.0.1
frozenlist==1.6.2
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.67.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
ldap3==2.9.1
MarkupSafe==2.1.5
multidict==6.4.4
narwhals==1.41.0
NEMO==6.0.2
numpy==2.3.0
openpyxl==3.1.4
packaging==24.2
pandas==2.3.0
pillow==10.3.0
platformdirs==4.2.2
propcache==0.3.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.0
pycparser==2.22
pydeck==0.9.1
pymilvus==2.5.10
pymodbus==3.3.2
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-dateutil==2.9.0
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.36.2
reportlab==4.2.2
requests==2.32.3
rpds-py==0.25.1
setuptools==69.5.1
six==1.16.0
smmap==5.0.2
sqlparse==0.5.0
streamlit==1.45.1
temporalio==1.12.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
types-protobuf==6.30.2.20250516
typing_extensions==4.13.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
virtualenv==20.26.2
watchdog==6.0.0
Werkzeug==3.0.4
wheel==0.43.0
yarl==1.20.0
